restructure (cont.)
njit functions doesn't work if there is class instances (self, cls) being used inside the function. So moving the bottom level functions outside of class, only leaving one top-level function in class to be overriden by subclasses
